Add files via upload
I developed a Face Attendance System using machine learning and Python as the core programming language. The system uses OpenCV and Mediapipe for real-time face detection and recognition, with NumPy for facial data processing. The project is built with a Flask backend and a responsive frontend using HTML, CSS, and JavaScript.

Attendance data is stored securely in an SQL database, allowing for easy retrieval and management. This system automates attendance tracking through a webcam.